Move copying of config later to avoid full dependency and asset recompilation
Following the current example, I noticed that every time I update runtime.exs or any other files in the config directory it leads to full recompilation of all dependencies. On a large project, it's a long step. So I recommend moving it down a bit so that we only use it when we do the actual compilation.