Refer to master key file instead of `rails secret`
`bin/rails secret` generates a `secret_key_base` value rather than a
master key.  By default, Rails will generate a master key in
`config/master.key` (or in e.g. `config/credentials/production.key` if
using environment-specific credentials), so we can refer to that
instead.  Note that the master key file is automatically added to
`.gitignore`, so it may not be present in a cloned repo.

Although `$(< config/master.key)` would be more idiomatic, it might be
less familiar to readers, so `$(cat config/master.key)` is used instead.