Avoid confusion from `nginxproxy.fly.dev` not being a real address
Currently this doc is confusing as it sounds like `nginxproxy.fly.dev` is a real domain name that your CNAME should point to ... but it's really talking about `<yourapp>.fly.dev`. I was confused by this and set up the wrong CNAME at first, and it looks like I'm [not the only one](https://community.fly.io/t/ssl-termination-error-with-custom-domain-cname-and-new-certificate/6754/4).