Fix typos and improve formatting in JS guide (#1079)
* Change "self contained" to "self-contained"

I'm not a native english speaker and apologize if both versions are correct. But from a quick google search it appears that "self-contained" is more common. Also, "self-contained" already occurs twice in this repo while "self contained" does not.

* Format file name as code for consistency

Other mentions of `package.json` further down on the page format the file name as code.

* Fix typos

* One cap

---------

Co-authored-by: catflydotio <95245363+catflydotio@users.noreply.github.com>