Minor update to Postgres snapshot restore (#1314)
* Minor update to Postgres snapshot restore

Adding a minor update to the instructions to restore a Postgres
snapshot. There was a post in the community forum from an user
getting an error creating a new Postgres app by restoring from
a snapshot. (https://community.fly.io/t/restoring-postgres-snapshot-fails-restore-volume-size-must-be-at-least-10gb/17937)
The problem was the size for the provisioned volume
for the new Postgres app being smaller than the original volume.

* Update postgres/managing/backup-and-restore.html.markerb

Co-authored-by: Andrea Anderson <andrea@fly.io>

---------

Co-authored-by: Andrés Schiavo <andres@fly.io>
Co-authored-by: Andrea Anderson <andrea@fly.io>