Deep Dive updates:
---

The front page had a number of (in my opinion) disjoint sentences that could leave the wrong impression.

One way to interpret the way it was:
- It normally will take a minimum of an afternoon, expect surprises, but we will partner with you.

I reworded it with the intent to convey:
- It normally will take a minimum of an afternoon, we will complete it in minutes.
- We have taken steps to minimize surprises, but if you encounter one we will partner with you.

I also added a link to public networking, and mentioned preconfigured databases.

---

I added Django in (it should be complete by the time this page goes live).  I removed repetitive sentences
and replaced them with a simple link.