Update fly_postgres_import.md
The motivation is to clarify a bit about what we can use as a source-uri.

I took hours to discover that I couldn't use localhost as a target in the URI, which makes total sense now that I know :) 

I'm still not convinced this is a good contribution, mainly because the copy is probably not abstract enough. But well, here it is.